requirements.txt
asgiref==3.3.1
Django==3.1.6
django-cors-headers==3.7.0
django-rest-framework==0.1.0
django-rest-multiple-models==2.1.3
django-webpack-loader==0.7.0
djangorestframework==3.12.2
gunicorn==20.0.4
importlib-metadata==3.4.0
Markdown==3.3.3
pytz==2021.1
shortuuid==1.0.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
whitenoise==5.2.0
zipp==3.4.0
